Rename: 커뮤니티 예외 위치 수정 커뮤니티에서 사용되는 예외는 공통으로 사용되므로 위치를 이동하였고 그에 따른 import 들을 수정하였다.
Related to: #37
