+++ src/main/java/pulleydoreurae/careerquestbackend/basiccommunity/service/CommonCommunityService.java
@@ -18,9 +18,9 @@ import pulleydoreurae.careerquestbackend.auth.repository.UserAccountRepository;
 import pulleydoreurae.careerquestbackend.basiccommunity.domain.dto.request.PostRequest;
 import pulleydoreurae.careerquestbackend.basiccommunity.domain.dto.response.PostResponse;
 import pulleydoreurae.careerquestbackend.basiccommunity.domain.entity.BasicPost;
-import pulleydoreurae.careerquestbackend.basiccommunity.exception.CommentNotFoundException;
-import pulleydoreurae.careerquestbackend.basiccommunity.exception.PostLikeNotFoundException;
-import pulleydoreurae.careerquestbackend.basiccommunity.exception.PostNotFoundException;
+import pulleydoreurae.careerquestbackend.common.community.exception.CommentNotFoundException;
+import pulleydoreurae.careerquestbackend.common.community.exception.PostLikeNotFoundException;
+import pulleydoreurae.careerquestbackend.common.community.exception.PostNotFoundException;
 import pulleydoreurae.careerquestbackend.common.community.repository.CommentRepository;
 import pulleydoreurae.careerquestbackend.common.community.repository.PostImageRepository;
 import pulleydoreurae.careerquestbackend.common.community.repository.PostLikeRepository;
