+++ src/main/java/pulleydoreurae/careerquestbackend/common/community/exception/CommentNotFoundException.java
@@ -1,4 +1,4 @@
-package pulleydoreurae.careerquestbackend.basiccommunity.exception;
+package pulleydoreurae.careerquestbackend.common.community.exception;
 
 /**
  * 댓글을 찾지 못했을때 발생하는 예외
