+++ src/main/java/pulleydoreurae/careerquestbackend/common/community/exception/PostSaveException.java
@@ -1,4 +1,4 @@
-package pulleydoreurae.careerquestbackend.basiccommunity.exception;
+package pulleydoreurae.careerquestbackend.common.community.exception;
 
 /**
  * 게시글 저장에 실패했을때 발생하는 예외
