+++ src/test/java/pulleydoreurae/careerquestbackend/basiccommunity/service/CommentServiceTest.java
@@ -24,8 +24,8 @@ import pulleydoreurae.careerquestbackend.basiccommunity.domain.dto.request.Comme
 import pulleydoreurae.careerquestbackend.basiccommunity.domain.dto.response.CommentResponse;
 import pulleydoreurae.careerquestbackend.basiccommunity.domain.entity.BasicComment;
 import pulleydoreurae.careerquestbackend.basiccommunity.domain.entity.BasicPost;
-import pulleydoreurae.careerquestbackend.basiccommunity.exception.CommentNotFoundException;
-import pulleydoreurae.careerquestbackend.basiccommunity.exception.PostNotFoundException;
+import pulleydoreurae.careerquestbackend.common.community.exception.CommentNotFoundException;
+import pulleydoreurae.careerquestbackend.common.community.exception.PostNotFoundException;
 import pulleydoreurae.careerquestbackend.common.community.repository.CommentRepository;
 import pulleydoreurae.careerquestbackend.common.community.domain.entity.Comment;
 import pulleydoreurae.careerquestbackend.common.community.domain.entity.Post;
