+++ src/test/java/pulleydoreurae/careerquestbackend/basiccommunity/service/CommonCommunityServiceTest.java
@@ -27,9 +27,9 @@ import pulleydoreurae.careerquestbackend.basiccommunity.domain.entity.BasicPost;
 import pulleydoreurae.careerquestbackend.basiccommunity.domain.entity.BasicPostImage;
 import pulleydoreurae.careerquestbackend.basiccommunity.domain.entity.BasicPostLike;
 import pulleydoreurae.careerquestbackend.basiccommunity.domain.entity.BasicPostViewCheck;
-import pulleydoreurae.careerquestbackend.basiccommunity.exception.CommentNotFoundException;
-import pulleydoreurae.careerquestbackend.basiccommunity.exception.PostLikeNotFoundException;
-import pulleydoreurae.careerquestbackend.basiccommunity.exception.PostNotFoundException;
+import pulleydoreurae.careerquestbackend.common.community.exception.CommentNotFoundException;
+import pulleydoreurae.careerquestbackend.common.community.exception.PostLikeNotFoundException;
+import pulleydoreurae.careerquestbackend.common.community.exception.PostNotFoundException;
 import pulleydoreurae.careerquestbackend.common.community.repository.CommentRepository;
 import pulleydoreurae.careerquestbackend.common.community.repository.PostImageRepository;
 import pulleydoreurae.careerquestbackend.common.community.repository.PostLikeRepository;
