+++ src/test/java/pulleydoreurae/careerquestbackend/basiccommunity/service/PostLikeServiceTest.java
@@ -23,7 +23,7 @@ import pulleydoreurae.careerquestbackend.basiccommunity.domain.dto.request.PostL
 import pulleydoreurae.careerquestbackend.basiccommunity.domain.dto.response.PostResponse;
 import pulleydoreurae.careerquestbackend.basiccommunity.domain.entity.BasicPost;
 import pulleydoreurae.careerquestbackend.basiccommunity.domain.entity.BasicPostLike;
-import pulleydoreurae.careerquestbackend.basiccommunity.exception.PostNotFoundException;
+import pulleydoreurae.careerquestbackend.common.community.exception.PostNotFoundException;
 import pulleydoreurae.careerquestbackend.common.community.repository.PostLikeRepository;
 import pulleydoreurae.careerquestbackend.common.community.domain.entity.Post;
 import pulleydoreurae.careerquestbackend.common.community.domain.entity.PostLike;
